+++ crates/rtlola-frontend/CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - TBD
+## [0.4.2] - 2021-05-27
+
+### Added
+- A trigger can now be annotated with streams. Their values are incorporated into the trigger message.
+
+## [0.4.1] - 2021-05-24
+
+### Changed
+- Minor changes to the MIR to be compatible with the interpreter.
+
+## [0.4.0] - 2021-04-08
 
 ### Changed
 - Complete revision of frontend